FastCore Refactor
This commit refactors the Mzinga.Shared project into a new Mzinga library project. This new project is based on the simpler (and faster) architecture design of MzingaCpp.

This commit also fixes up all of the other projects so they work with
the new Mzinga dll.

The end result is a simpler codebase that's easier to maintain, builds faster and runs faster.