Trainer: Puzzle candidates are now validated as actual puzzles
For the purposes of Mzinga, a "puzzle" is a game position whereby the
current player has exactly one valid move that will let them win the game
in a given number of plies.

So a one-ply puzzle means there is exactly one move to immediately win the
game. Nothing special, but a baseline.

A two-ply puzzle means there is exactly one move that will force the
opponent to commit suicide, as all of their moves will kill them.

A higher-ply puzzle means there is exactly one move to start forcing a
win in that many plies, though there may be multiple branches to get
there, in later plies. So the total number of solutions may be greater
than 1, but all will start from the same move. Odd-plies mean the player
will force the win, while even-plies mean the player will force the
opponent to commit suicide.

This long-awaited work required several changes, namely to stop
reporting every possible "candidate" while the trainer is running, as
the vast majority were false-positives. While the qualifications for
being a candidate have broadened, now the candidates are queued up until
after a session (battle or auto train), then validated. Only validated
puzzles are reported to the console. There is now a PuzzleCandidate in
Core which does this validation, as well as tests on known good puzzles.

In addition, the GameAI tests have been updated to also validate that
Mzinga is able to solve all of these puzzles. This is probably the most
useful addition to the tests, as it provides a concrete mechanism for
validating that the AI always picks the absolute best move if there is
one.