Viewer: Added options for free panning and zoom of game board
New ViewerConfig options to disable auto centering and auto zoom of the
board. Exposed in the config window and the top menu. Added a new
context menu to the game board to force a manual recenter / zoom.

When the options are disabled, user can freely pan the board by clicking
and dragging, and zoom by using the mouse wheel or pinch to zoom
gestures. The board can also now be interacted with even while the
engine is busy, allowing for adjusting the board's position and zoom
even while the AI is thinking.

Closes #111