Trainer: AutoTrain now supports using MzingaAutoTrainConfig.xml
The previous workflow required creating a profile for MzingaTrainer to
AutoTrain. Now, if no profile is offered, MzingaTrainer will use the
current default engine config as the starting point for training.

Furthermore, it will save its intermediate results to an
`MzingaAutoTrainConfig.xml` file, which will also be detected and loaded
when no profile is provided. This will make it easier to get started
with autotraining.

Also, the `exportai` command has been updated to detect the presense of
the new config file, and instead export profiles clearly labeled as
coming from AutoTrain, with set IDs that won't conflict with profiles
created based on an actual released version. This will make it easier to
build up and test a set of profiles for trainer battles.